Merge layout into single section
Previous changed are reversed due to issues with sharing BinomialTree
and Analyzer objects between tabs. As a result of this change whole
layout specification is now contained in single file layout.py.
Layout of the app is redesigned to fit single page.

The Cytoscape is now labelled with stock prices. Nodes multiselection
is wired with querying matrix of prices (tree of prices).

Default parameters values are updated and specified with respect to year.
The adjust_params function from utils module is responsible for
transforming them into appropriate scale before usage. All user defined
parameters must be specified with respect to year to avoid confusions.