Draw plots for chosen path
Layout is redesigned again. Number of parameters is reduced. Submission
button are removed. Barrier options are temporarily not supported.

Functions from tree_paths module are improved and now support extracting
paths with indices of prices in the tree. This functionality is essential
when it comes to efficient communication between Cytoscape object and
BinomialTree and its properties, while choosing the path.